+++ src/components/AddPaper/Contributions/Contributions.js
@@ -1,25 +1,25 @@
-import { faAngleDown, faExclamationTriangle, faMagic, faFlask } from '@fortawesome/free-solid-svg-icons';
+import env from '@beam-australia/react-env';
+import { faAngleDown, faExclamationTriangle, faFlask, faMagic } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import Tippy from '@tippyjs/react';
 import Abstract from 'components/AddPaper/Abstract/Abstract';
 import AbstractModal from 'components/AddPaper/AbstractModal/AbstractModal';
+import BioAssaysModal from 'components/AddPaper/BioAssaysModal/BioAssaysModal';
 import EntityRecognition from 'components/AddPaper/EntityRecognition/EntityRecognition';
-import useDetermineResearchField from 'components/AddPaper/EntityRecognition/useDetermineResearchField';
+import useBioassays from 'components/AddPaper/hooks/useBioassays';
 import useEntityRecognition from 'components/AddPaper/hooks/useEntityRecognition';
 import useFeedbacks from 'components/AddPaper/hooks/useFeedbacks';
-import useBioassays from 'components/AddPaper/hooks/useBioassays';
 import Confirm from 'components/Confirmation/Confirmation';
 import AddContributionButton from 'components/ContributionTabs/AddContributionButton';
 import ContributionTab from 'components/ContributionTabs/ContributionTab';
-import { StyledContributionTabs, GlobalStyle } from 'components/ContributionTabs/styled';
 import StatementBrowser from 'components/StatementBrowser/StatementBrowser';
 import Tooltip from 'components/Utils/Tooltip';
-import Tabs from 'rc-tabs';
+import { BIOASSAYS_FIELDS_LIST } from 'constants/nlpFieldLists';
+import Tabs from 'components/Tabs/Tabs';
 import { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import BIOASSAYS_FIELDS_LIST from 'constants/bioassayFieldList';
-import BioAssaysModal from 'components/AddPaper/BioAssaysModal/BioAssaysModal';
 import { Button, Col, Row, UncontrolledAlert } from 'reactstrap';
+import { determineActiveNERService } from 'services/orkgNlp/index';
 import {
     createContributionAction as createContribution,
     deleteContributionAction as deleteContribution,
@@ -32,7 +32,6 @@ import {
     updateContributionLabelAction as updateContributionLabel,
 } from 'slices/addPaperSlice';
 import { updateSettings } from 'slices/statementBrowserSlice';
-import env from '@beam-australia/react-env';
 import ContributionsHelpTour from './ContributionsHelpTour';
 
 const Contributions = () => {
@@ -50,13 +49,12 @@ const Contributions = () => {
         abstract,
     } = useSelector(state => state.addPaper);
     const [isOpenAbstractModal, setIsOpenAbstractModal] = useState(false);
-    const [isComputerScienceField, setIsComputerScienceField] = useState(false);
+    const [activeNERService, setActiveNERService] = useState(null);
     const { resources, properties, values } = useSelector(state => state.statementBrowser);
-    const { determineField } = useDetermineResearchField();
 
     const isBioassayField = BIOASSAYS_FIELDS_LIST.includes(selectedResearchField);
 
-    const { handleSaveFeedback } = useEntityRecognition({ isComputerScienceField });
+    const { handleSaveFeedback } = useEntityRecognition({ activeNERService });
     const { handleSaveBioassaysFeedback } = useBioassays();
     const { handleSavePredicatesRecommendationFeedback } = useFeedbacks();
 
@@ -65,8 +63,8 @@ const Contributions = () => {
     const dispatch = useDispatch();
 
     useEffect(() => {
-        (async () => setIsComputerScienceField(await determineField({ field: selectedResearchField })))();
-    }, [determineField, selectedResearchField]);
+        (async () => setActiveNERService(await determineActiveNERService(selectedResearchField)))();
+    }, [selectedResearchField]);
 
     useEffect(() => {
         // if there is no contribution yet, create the first one
@@ -87,7 +85,7 @@ const Contributions = () => {
     }, [contributions.allIds.length, dispatch, selectedResearchField]);
 
     const handleNextClick = async () => {
-        if (isComputerScienceField) {
+        if (activeNERService) {
             handleSaveFeedback();
         }
         if (isBioassayField) {
@@ -149,7 +147,7 @@ const Contributions = () => {
         dispatch(openTour(step));
     };
 
-    const showAbstractWarning = isComputerScienceField && !abstract;
+    const showAbstractWarning = activeNERService && !abstract;
     const onTabChange = key => {
         handleSelectContribution(key);
     };
@@ -162,7 +160,6 @@ const Contributions = () => {
 
     return (
         <div>
-            <GlobalStyle />
             <div className="d-flex align-items-center mt-4 mb-4">
                 <h2 className="h4 flex-shrink-0">
                     <Tooltip
@@ -196,7 +193,7 @@ const Contributions = () => {
                             <Icon icon={faFlask} /> Add Bioassay
                         </Button>
                     )}
-                    {!isComputerScienceField ? (
+                    {!activeNERService ? (
                         <Button onClick={() => dispatch(toggleAbstractDialog())} outline size="sm" color="smart">
                             {!showAbstractWarning ? <Icon icon={faMagic} /> : <Icon icon={faExclamationTriangle} className="text-warning" />} Abstract
                             annotator
@@ -227,49 +224,47 @@ const Contributions = () => {
             )}
             <Row className="mt-2 g-0">
                 <Col md="9">
-                    <StyledContributionTabs>
-                        <Tabs
-                            renderTabBar={renderTabBar}
-                            tabBarExtraContent={<AddContributionButton onClick={() => dispatch(createContribution({ selectAfterCreation: true }))} />}
-                            moreIcon={<Icon size="lg" icon={faAngleDown} />}
-                            activeKey={selectedContribution}
-                            onChange={onTabChange}
-                            destroyInactiveTabPane={true}
-                            items={contributions.allIds.map(contributionId => {
-                                const contribution = contributions.byId[contributionId];
-                                return {
-                                    label: (
-                                        <ContributionTab
-                                            handleChangeContributionLabel={handleChange}
-                                            isSelected={contribution.id === selectedContribution}
-                                            canDelete={contributions.allIds.length !== 1}
-                                            contribution={contribution}
-                                            key={contribution.id}
-                                            toggleDeleteContribution={toggleDeleteContribution}
+                    <Tabs
+                        renderTabBar={renderTabBar}
+                        tabBarExtraContent={<AddContributionButton onClick={() => dispatch(createContribution({ selectAfterCreation: true }))} />}
+                        moreIcon={<Icon size="lg" icon={faAngleDown} />}
+                        activeKey={selectedContribution}
+                        onChange={onTabChange}
+                        destroyInactiveTabPane={true}
+                        items={contributions.allIds.map(contributionId => {
+                            const contribution = contributions.byId[contributionId];
+                            return {
+                                label: (
+                                    <ContributionTab
+                                        handleChangeContributionLabel={handleChange}
+                                        isSelected={contribution.id === selectedContribution}
+                                        canDelete={contributions.allIds.length !== 1}
+                                        contribution={contribution}
+                                        key={contribution.id}
+                                        toggleDeleteContribution={toggleDeleteContribution}
+                                        enableEdit={true}
+                                    />
+                                ),
+                                key: contribution.id,
+                                children: (
+                                    <div className="contributionData p-4">
+                                        <StatementBrowser
                                             enableEdit={true}
+                                            syncBackend={false}
+                                            openExistingResourcesInDialog={false}
+                                            initialSubjectId={contribution.resourceId}
+                                            initialSubjectLabel={contribution.label}
+                                            renderTemplateBox={true}
                                         />
-                                    ),
-                                    key: contribution.id,
-                                    children: (
-                                        <div className="contributionData">
-                                            <StatementBrowser
-                                                enableEdit={true}
-                                                syncBackend={false}
-                                                openExistingResourcesInDialog={false}
-                                                initialSubjectId={contribution.resourceId}
-                                                initialSubjectLabel={contribution.label}
-                                                renderTemplateBox={true}
-                                            />
-                                        </div>
-                                    ),
-                                };
-                            })}
-                        />
-                    </StyledContributionTabs>
+                                    </div>
+                                ),
+                            };
+                        })}
+                    />
                 </Col>
 
                 <Col lg="3" className="ps-lg-3 mt-2">
-                    <EntityRecognition isComputerScienceField={isComputerScienceField} />
+                    <EntityRecognition activeNERService={activeNERService} />
                 </Col>
             </Row>