+++ src/components/StatementBrowser/DatatypeSelector/DatatypeSelector.js
@@ -1,11 +1,36 @@
 import { useContext, useCallback } from 'react';
 import Select, { components } from 'react-select';
-// import { useSelector } from 'react-redux';
 import DATA_TYPES, { getConfigByType } from 'constants/DataTypes';
+import { ENTITIES } from 'constants/graphSettings';
+import ConditionalWrapper from 'components/Utils/ConditionalWrapper';
 import { ThemeContext } from 'styled-components';
 import { SelectGlobalStyle } from 'components/Autocomplete/styled';
 import Tippy from '@tippyjs/react';
 import PropTypes from 'prop-types';
+import { Link } from 'react-router-dom';
+import { reverse } from 'named-urls';
+import ROUTES from 'constants/routes.js';
+
+const TypeTooltipContent = ({ valueClass, entity, switchEntityType }) => (
+    <>
+        {valueClass || !switchEntityType ? 'Type is determined by the template.' : 'Changing the type is not possible'}
+        {valueClass && entity === ENTITIES.RESOURCE && (
+            <div>
+                Only instances of{' '}
+                <Link target="_blank" to={reverse(ROUTES.CLASS, { id: valueClass.id })}>
+                    {valueClass.label}
+                </Link>{' '}
+                are valid.
+            </div>
+        )}
+    </>
+);
+
+TypeTooltipContent.propTypes = {
+    entity: PropTypes.string,
+    valueClass: PropTypes.object,
+    switchEntityType: PropTypes.bool,
+};
 
 const DatatypeSelector = props => {
     // const isCurationAllowed = useSelector(state => state.auth.user?.isCurationAllowed);
@@ -79,19 +104,38 @@ const DatatypeSelector = props => {
 
     return (
         <>
-            <Select
-                styles={customStyles}
-                classNamePrefix="react-select-dark"
-                value={getConfigByType(props.valueType)}
-                components={{ Option: CustomOption }}
-                options={!props.entity ? DATA_TYPES : DATA_TYPES.filter(dt => dt._class === props.entity)}
-                onChange={v => props.setValueType(v.type)}
-                getOptionValue={({ type }) => type}
-                getOptionLabel={({ name }) => name}
-                isClearable={false}
-                menuPortalTarget={props.menuPortalTarget}
-                inputId="datatypeSelector"
-            />
+            <ConditionalWrapper
+                condition={props.isDisabled}
+                wrapper={children => (
+                    <Tippy
+                        interactive
+                        content={
+                            <TypeTooltipContent
+                                switchEntityType={props.entity && DATA_TYPES.filter(dt => dt._class === props.entity).length <= 1}
+                                entity={getConfigByType(props.valueType)._class}
+                                valueClass={props.valueClass}
+                            />
+                        }
+                    >
+                        <span>{children}</span>
+                    </Tippy>
+                )}
+            >
+                <Select
+                    styles={customStyles}
+                    classNamePrefix="react-select-dark"
+                    value={getConfigByType(props.valueType)}
+                    components={{ Option: CustomOption }}
+                    options={!props.entity ? DATA_TYPES : DATA_TYPES.filter(dt => dt._class === props.entity)}
+                    onChange={v => props.setValueType(v.type)}
+                    getOptionValue={({ type }) => type}
+                    getOptionLabel={({ name }) => name}
+                    isClearable={false}
+                    menuPortalTarget={props.menuPortalTarget}
+                    inputId="datatypeSelector"
+                    isDisabled={props.isDisabled}
+                />
+            </ConditionalWrapper>
             <SelectGlobalStyle />
         </>
     );
@@ -104,12 +148,16 @@ DatatypeSelector.propTypes = {
     disableBorderRadiusLeft: PropTypes.bool,
     disableBorderRadiusRight: PropTypes.bool,
     menuPortalTarget: PropTypes.object,
+    isDisabled: PropTypes.bool,
+    valueClass: PropTypes.object,
 };
 
 DatatypeSelector.defaultProps = {
     disableBorderRadiusLeft: false,
     disableBorderRadiusRight: true,
     menuPortalTarget: null,
+    isDisabled: false,
+    valueClass: null,
 };
 
 export default DatatypeSelector;