+++ src/components/UserProfile/TabsContainer.js
@@ -1,12 +1,11 @@
 /* eslint-disable linebreak-style */
-import { useState } from 'react';
-import Tabs from 'rc-tabs';
-import { GlobalStyle, StyledContributionTabs } from 'components/ContributionTabs/styled';
-import { Container } from 'reactstrap';
-import { CLASSES } from 'constants/graphSettings';
+import Tabs from 'components/Tabs/Tabs';
 import Items from 'components/UserProfile/Items';
-import { useParams } from 'react-router-dom';
+import { CLASSES } from 'constants/graphSettings';
 import PropTypes from 'prop-types';
+import { useState } from 'react';
+import { useParams } from 'react-router-dom';
+import { Container } from 'reactstrap';
 
 function TabsContainer({ currentUserId }) {
     const params = useParams();
@@ -23,55 +22,52 @@ function TabsContainer({ currentUserId }) {
 
     return (
         <Container className="mt-2 p-0">
-            <GlobalStyle />
-            <StyledContributionTabs disablePadding={true}>
-                <Tabs
-                    destroyInactiveTabPane={true}
-                    onChange={onTabChange}
-                    activeKey={isActiveTab ?? '1'}
-                    items={[
-                        {
-                            label: 'Comparisons',
-                            key: '1',
-                            children: (
-                                <Container className="p-0">
-                                    <Items filterLabel="comparisons" filterClass={CLASSES.COMPARISON} userId={userId} />
-                                </Container>
-                            ),
-                        },
+            <Tabs
+                destroyInactiveTabPane={true}
+                onChange={onTabChange}
+                activeKey={isActiveTab ?? '1'}
+                items={[
+                    {
+                        label: 'Comparisons',
+                        key: '1',
+                        children: (
+                            <Container className="p-0">
+                                <Items filterLabel="comparisons" filterClass={CLASSES.COMPARISON} userId={userId} />
+                            </Container>
+                        ),
+                    },
 
-                        {
-                            label: 'Papers',
-                            key: '2',
-                            children: (
-                                <Container className="p-0">
-                                    <Items filterLabel="papers" filterClass={CLASSES.PAPER} userId={userId} showDelete={userId === currentUserId} />
-                                </Container>
-                            ),
-                        },
-                        {
-                            label: 'Templates',
-                            key: '3',
-                            children: (
-                                <Container className="p-0">
-                                    <Items filterLabel="templates" filterClass={CLASSES.TEMPLATE} userId={userId} showDelete={false} />
-                                </Container>
-                            ),
-                        },
+                    {
+                        label: 'Papers',
+                        key: '2',
+                        children: (
+                            <Container className="p-0">
+                                <Items filterLabel="papers" filterClass={CLASSES.PAPER} userId={userId} showDelete={userId === currentUserId} />
+                            </Container>
+                        ),
+                    },
+                    {
+                        label: 'Templates',
+                        key: '3',
+                        children: (
+                            <Container className="p-0">
+                                <Items filterLabel="templates" filterClass={CLASSES.TEMPLATE} userId={userId} showDelete={false} />
+                            </Container>
+                        ),
+                    },
 
-                        {
-                            label: 'Reviews',
-                            key: '4',
+                    {
+                        label: 'Reviews',
+                        key: '4',
 
-                            children: (
-                                <Container className="p-0">
-                                    <Items filterLabel="reviews" filterClass={CLASSES.SMART_REVIEW_PUBLISHED} userId={userId} showDelete={false} />
-                                </Container>
-                            ),
-                        },
-                    ]}
-                />
-            </StyledContributionTabs>
+                        children: (
+                            <Container className="p-0">
+                                <Items filterLabel="reviews" filterClass={CLASSES.SMART_REVIEW_PUBLISHED} userId={userId} showDelete={false} />
+                            </Container>
+                        ),
+                    },
+                ]}
+            />
         </Container>
     );
 }