+++ src/services/backend/benchmarks.js
@@ -7,5 +7,9 @@ export const benchmarksUrl = `${url}benchmarks/`;
 // https://gitlab.com/TIBHannover/orkg/orkg-backend/-/issues/263
 
 // Get a set of benchmarks where each benchmark item is a summary with the following attributes:
-// research_problem, total_papers, total_datasets, total_codes
+// research_field, total_papers, total_datasets, total_codes
 export const getBenchmarksByResearchFieldId = rfId => submitGetRequest(`${benchmarksUrl}summary/research-field/${encodeURIComponent(rfId)}`);
+
+// Get a set of benchmarks where each benchmark item is a summary with the following attributes:
+// research_field, research_problem, total_papers, total_datasets, total_codes
+export const getAllBenchmarks = () => submitGetRequest(`${benchmarksUrl}summary/`);